fixed the array mutators
Got the buttons to push in the correct values and made it so that it won't push until next is pressed, this way the user can't spam every "which level did you hear it" which will then push every value. This way, only one value is pushed in. Will look at graph next and I would consider using some sort of radio button type for the "lowest level" question.